requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.2.2
    # via ortools
annotated-types==0.7.0
    # via pydantic
antares-craft==0.2.9
    # via -r requirements.in
antares-study-version==1.0.15
    # via antares-craft
antares-timeseries-generation==0.1.7
    # via antares-craft
antlr4-python3-runtime==4.13.1
    # via -r requirements.in
anytree==2.12.1
    # via -r requirements.in
bottleneck==1.4.2
    # via linopy
certifi==2025.1.31
    # via
    #   netcdf4
    #   pyogrio
    #   pyproj
    #   requests
cftime==1.6.4.post1
    # via netcdf4
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   antares-study-version
    #   dask
cloudpickle==3.1.1
    # via dask
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dask==2025.4.0
    # via linopy
deprecation==2.1.0
    # via
    #   linopy
    #   pypsa
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via dask
geopandas==1.0.1
    # via pypsa
highspy==1.10.0
    # via pypsa
idna==3.10
    # via requests
immutabledict==4.2.1
    # via ortools
importlib-metadata==8.6.1
    # via dask
kiwisolver==1.4.8
    # via matplotlib
linopy==0.5.3
    # via pypsa
locket==1.0.0
    # via partd
matplotlib==3.10.1
    # via
    #   pypsa
    #   seaborn
netcdf4==1.7.2
    # via pypsa
networkx==3.4.2
    # via pypsa
numexpr==2.10.2
    # via linopy
numpy==1.26.4
    # via
    #   -r requirements.in
    #   antares-craft
    #   antares-timeseries-generation
    #   bottleneck
    #   cftime
    #   contourpy
    #   geopandas
    #   highspy
    #   linopy
    #   matplotlib
    #   netcdf4
    #   numexpr
    #   ortools
    #   pandas
    #   pyogrio
    #   pypsa
    #   scipy
    #   seaborn
    #   shapely
    #   xarray
ortools==9.9.3963
    # via -r requirements.in
packaging==25.0
    # via
    #   dask
    #   deprecation
    #   geopandas
    #   matplotlib
    #   pyogrio
    #   xarray
pandas==2.2.3
    # via
    #   antares-craft
    #   antares-study-version
    #   antares-timeseries-generation
    #   geopandas
    #   ortools
    #   pypsa
    #   seaborn
    #   xarray
partd==1.4.2
    # via dask
pillow==11.2.1
    # via matplotlib
polars==1.27.1
    # via linopy
protobuf==6.30.2
    # via ortools
psutil==7.0.0
    # via antares-craft
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   antares-craft
pydantic-core==2.33.2
    # via pydantic
pyogrio==0.10.0
    # via geopandas
pyparsing==3.2.3
    # via matplotlib
pyproj==3.7.1
    # via geopandas
pypsa==0.34.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   antares-timeseries-generation
    #   dask
requests==2.32.3
    # via antares-craft
scipy==1.10.1
    # via
    #   -r requirements.in
    #   linopy
    #   pypsa
seaborn==0.13.2
    # via pypsa
shapely==2.0.7
    # via
    #   geopandas
    #   pypsa
six==1.17.0
    # via
    #   anytree
    #   python-dateutil
toolz==1.0.0
    # via
    #   dask
    #   linopy
    #   partd
tqdm==4.67.1
    # via linopy
typing-extensions==4.13.2
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via requests
validators==0.34.0
    # via pypsa
xarray==2025.3.1
    # via
    #   linopy
    #   pypsa
zipp==3.21.0
    # via importlib-metadata
